% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSpeciesColumn.R
\name{AddSpeciesColumn}
\alias{AddSpeciesColumn}
\title{Add Species Column}
\usage{
AddSpeciesColumn(data, taxon_col = "nom", level_col = "Determination")
}
\arguments{
\item{data}{A data frame containing taxonomic data.}

\item{taxon_col}{Character. Column name for the taxon. Default is \code{"nom"}.}

\item{level_col}{Character. Column name for the determination level. Default is \code{"Determination"}.}
}
\value{
The modified data frame with a new \code{Species} column.
}
\description{
Convenience function to extract and append a species column from the dataset.
}
\examples{
Dat <- data.frame(nom = c("A", "B"), Determination = c("Species", "Genus"))
AddSpeciesColumn(Dat)
}
