% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT_SensitivityAnalysis.R
\name{HT_SensitivityAnalysis}
\alias{HT_SensitivityAnalysis}
\title{Sensitivity Analysis for Taxonomic Harmonization}
\usage{
HT_SensitivityAnalysis(
  InvData,
  HTs = seq(0.05, 0.95, by = 0.05),
  abundance = FALSE,
  taxonomic_hierarchy,
  taxon_col = "nom",
  level_col = "Determination",
  count_col = "total",
  abundance_col = "abundance"
)
}
\arguments{
\item{InvData}{A data frame containing the taxonomic data.}

\item{HTs}{Numeric vector of homogenization thresholds to test. Default is \code{seq(0.05, 0.95, by = 0.05)}.}

\item{abundance}{Logical. If \code{TRUE}, calculates lost abundance or another ecological metric in addition to lost count information. Default is \code{FALSE}.}

\item{taxonomic_hierarchy}{Character vector of the ordered taxonomic levels.}

\item{taxon_col}{Character. Column name for the taxon. Default is \code{"nom"}.}

\item{level_col}{Character. Column name for the determination level. Default is \code{"Determination"}.}

\item{count_col}{Character. Column name for the counts. Default is \code{"total"}.}

\item{abundance_col}{Character. Column name for abundance (used if \code{abundance = TRUE}). Default is \code{"abundance"}.}
}
\value{
A data frame containing richness, lost count, and lost abundance (or other ecological metric) for each threshold.
}
\description{
Performs logical tree analysis across multiple homogenization thresholds (HTs) to estimate the loss of counts (i.e. input for taxonomic harmonisation, could be occurrences, abundance etc.), abundance, and richness.
}
\examples{
# See generic package vignette for a full sensitivity analysis walk-through. TO-DO
}
